+++ src/components/common/Spacing/utils.tsx
@@ -2,14 +2,20 @@ import { DefaultTheme } from 'styled-components'
 
 import { Spacing } from '../../../types/Spacing'
 
+export const isSpacing = (
+	size?: Spacing | string | number
+): size is Spacing => {
+	return Spacing.includes(size as Spacing)
+}
+
 export const getSpacingCssValue = (
 	theme: DefaultTheme,
 	value?: Spacing | string | number,
 	defaultValue = '0'
 ): string => {
-	if (Spacing.includes(value as unknown as Spacing)) {
+	if (isSpacing(value)) {
 		// Spacing size from theme file
-		return `${theme.$pc.spacing[value as unknown as Spacing]}px`
+		return `${theme.$pc.spacing[value]}px`
 	} else if (typeof value === 'number') {
 		// Number type - implicitly px
 		return `${value}px`